Extracts countdown functionality to Countdown component
